added test and refactored env variables
