+++ rasa/__main__.py
@@ -8,7 +8,7 @@ from rasa.cli import scaffold, run, train, interactive, shell, test, visualize, 
 from rasa.cli.arguments.default_arguments import add_logging_options
 from rasa.cli.utils import parse_last_positional_argument_as_model_path
 from rasa.utils.common import set_log_level
-from rasa.utils.tensorflow import setup_tf_environment
+from rasa.utils.tensorflow.environment import setup_tf_environment
 
 logger = logging.getLogger(__name__)
